[HU-0009] Se crearon y reorganizaron carpetas para mejor organización de archivos de acuerdo a cada modulo.
